E2E Validation done and code refactoring
